0x0F Project 0, change sum from long back to int.

+++ 0x0F-variadic_functions/0-sum_them_all.c
@@ -9,7 +9,7 @@
 int sum_them_all(const unsigned int n, ...)
 {
 	unsigned int i;
-	long int sum;
+	int sum;
 
 	va_list ar;
